Add MinecraftServer class in api module for central handling of Bedrock and Java Edition
