+++ .coveragerc
@@ -747,6 +747,7 @@ omit =
     homeassistant/components/mill/climate.py
     homeassistant/components/mill/sensor.py
     homeassistant/components/minecraft_server/__init__.py
+    homeassistant/components/minecraft_server/api.py
     homeassistant/components/minecraft_server/binary_sensor.py
     homeassistant/components/minecraft_server/coordinator.py
     homeassistant/components/minecraft_server/entity.py
